requirements.txt
# Core dependencies
requests>=2.31.0          # For Samsara API calls
python-dotenv>=1.0.0      # For environment variables
pandas>=2.0.0             # For CSV processing
jinja2>=3.1.0            # For email templates

# Windows-specific for Outlook integration
pywin32>=305             # For Outlook COM automation

# Optional but recommended
openpyxl>=3.1.0          # If you need Excel support
python-dateutil>=2.8.0   # For date handling
pytz>=2023.3             # For timezone handling